+++ packages/insomnia/src/ui/components/panes/placeholder-request-pane.tsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import { ForceToWorkspace } from '../../../common/import';
 import { createRequest } from '../../hooks/create-request';
-import { importFile } from '../../redux/modules/import';
+import { importFile } from '../../import';
 import { selectActiveWorkspace, selectSettings } from '../../redux/selectors';
 import { Hotkey } from '../hotkey';
 import { Pane, PaneBody, PaneHeader } from './pane';
